Move Comment back to UML model
