+++ gaphor/SysML/tests/test_diagramitems.py
@@ -1,11 +1,11 @@
 import pytest
 
 from gaphor.core.modeling import Presentation
-from gaphor.diagram.general import CommentLineItem
 from gaphor.diagram.presentation import Classified, Named
 from gaphor.diagram.support import get_model_element
 from gaphor.SysML import diagramitems
 from gaphor.UML import Classifier, NamedElement, Relationship
+from gaphor.UML.general import CommentLineItem
 
 
 def _issubclass(child, parent):
