+++ gaphor/UML/diagramitems.py
@@ -3,11 +3,12 @@
 This module makes it easier to load a diagram item.
 """
 
-# ruff: noqa: F403
+# ruff: noqa: F401,F403
 
 from gaphor.UML.actions import *
 from gaphor.UML.classes import *
 from gaphor.UML.deployments import *
+from gaphor.UML.general import CommentItem, CommentLineItem
 from gaphor.UML.interactions import *
 from gaphor.UML.profiles import *
 from gaphor.UML.states import *
