+++ gaphor/UML/general/__init__.py
@@ -1,0 +1,3 @@
+from gaphor.UML.general import connectors as connectors
+from gaphor.UML.general.comment import CommentItem as CommentItem
+from gaphor.UML.general.comment import CommentLineItem as CommentLineItem
