+++ gaphor/UML/general/tests/test_comment.py
@@ -1,14 +1,12 @@
-"""Comment and comment line items connection adapters tests."""
+"""UML.Comment and comment line items connection adapters tests."""
 
 from typing import TypeVar
 
 from gaphor import UML
-from gaphor.core.modeling import Comment
-from gaphor.diagram.general.comment import CommentItem
-from gaphor.diagram.general.commentline import CommentLineItem
 from gaphor.diagram.tests.fixtures import allow, connect, disconnect
 from gaphor.UML.classes.generalization import GeneralizationItem
 from gaphor.UML.classes.klass import ClassItem
+from gaphor.UML.general.comment import CommentItem, CommentLineItem
 from gaphor.UML.usecases.actor import ActorItem
 
 T = TypeVar("T")
@@ -21,7 +19,7 @@ T = TypeVar("T")
 
 def test_commentline_annotated_element(create, diagram):
     """Test comment line item annotated element creation."""
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
 
     connect(line, line.head, comment)
@@ -33,7 +31,7 @@ def test_commentline_annotated_element(create, diagram):
 def test_commentline_same_comment_glue(create):
     """Test comment line item gluing to already connected comment item."""
 
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
 
     connect(line, line.head, comment)
@@ -43,7 +41,7 @@ def test_commentline_same_comment_glue(create):
 
 def test_commentline_element_connect(create, diagram):
     """Test comment line connecting to comment and actor items."""
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
     ac = create(ActorItem, UML.Actor)
 
@@ -64,7 +62,7 @@ def test_commentline_glie_to_item_with_no_subject(create, diagram):
 
 def test_commentline_item_with_no_subject_connect(create, diagram):
     """Test comment line connecting to comment and actor items."""
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
     gi = create(GeneralizationItem)
 
@@ -76,7 +74,7 @@ def test_commentline_item_with_no_subject_connect(create, diagram):
 
 def test_commentline_element_reconnect(create, diagram):
     """Test comment line connecting to comment and actor items."""
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
     ac = create(ActorItem, UML.Actor)
 
@@ -96,7 +94,7 @@ def test_commentline_element_reconnect(create, diagram):
 
 def test_commentline_element_disconnect(create, diagram):
     """Test comment line connecting to comment and disconnecting actor item."""
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
     ac = create(ActorItem, UML.Actor)
 
@@ -124,7 +122,7 @@ def test_commentline_relationship_disconnect(create, sanitizer_service):
     assert gen.subject
 
     # now, connect comment to a generalization (relationship)
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
     connect(line, line.head, comment)
     connect(line, line.tail, gen)
@@ -141,7 +139,7 @@ def test_commentline_relationship_disconnect(create, sanitizer_service):
 def test_commentline_unlink(create):
     """Test comment line unlinking."""
     clazz = create(ClassItem, UML.Class)
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
 
     connect(line, line.head, comment)
@@ -166,7 +164,7 @@ def test_commentline_unlink(create):
 def test_commentline_element_unlink(create, diagram):
     """Test comment line unlinking using a class item."""
     clazz = create(ClassItem, UML.Class)
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
 
     connect(line, line.head, comment)
@@ -203,7 +201,7 @@ def test_commentline_relationship_unlink(create, sanitizer_service):
     assert gen.subject
 
     # now, connect comment to a generalization (relationship)
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line = create(CommentLineItem)
     connect(line, line.head, comment)
     connect(line, line.tail, gen)
@@ -223,7 +221,7 @@ def test_commentline_linked_to_same_element_twice(create):
     clazz = create(ClassItem, UML.Class)
 
     # now, connect comment to a generalization (relationship)
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     line1 = create(CommentLineItem)
     connect(line1, line1.head, comment)
     connect(line1, line1.tail, clazz)
