+++ gaphor/UML/sanitizerservice.py
@@ -14,10 +14,10 @@ from gaphor.core.modeling.event import (
     DiagramUpdateRequested,
 )
 from gaphor.diagram.deletable import deletable
-from gaphor.diagram.general.diagramitems import CommentLineItem
 from gaphor.event import Notification
 from gaphor.i18n import gettext
 from gaphor.transaction import TransactionBegin, TransactionCommit, TransactionRollback
+from gaphor.UML.general.comment import CommentLineItem
 
 
 def undo_guard(func):
