+++ gaphor/core/modeling/__init__.py
@@ -2,7 +2,6 @@
 
 from gaphor.core.modeling.base import Base, Id, swap_element_type
 from gaphor.core.modeling.coremodel import (
-    Comment,
     ElementChange,
     PendingChange,
     Picture,
