+++ gaphor/core/modeling/element.py
@@ -15,12 +15,9 @@ from gaphor.core.modeling.properties import (
 )
 
 if TYPE_CHECKING:
-    from gaphor.core.modeling.coremodel import (
-        Comment,
-    )
     from gaphor.core.modeling.diagram import Diagram
     from gaphor.core.modeling.presentation import Presentation
-    from gaphor.UML import Relationship
+    from gaphor.UML import Comment, Relationship
 
 log = logging.getLogger(__name__)
