+++ gaphor/core/modeling/tests/test_modelinglanguage.py
@@ -14,7 +14,6 @@ def test_lookup_element():
 
     assert modeling_language.lookup_element("Diagram")
     assert modeling_language.lookup_element("Element")
-    assert modeling_language.lookup_element("Comment")
     assert modeling_language.lookup_element("Presentation")
     assert modeling_language.lookup_element("StyleSheet")
     assert not modeling_language.lookup_element("NonExistent")
