+++ gaphor/diagram/general/diagramitems.py
@@ -1,15 +1,11 @@
 # ruff: noqa: F401
 
-from gaphor.diagram.general.comment import CommentItem
-from gaphor.diagram.general.commentline import CommentLineItem
 from gaphor.diagram.general.diagramitem import DiagramItem
 from gaphor.diagram.general.metadata import MetadataItem
 from gaphor.diagram.general.picture import PictureItem
 from gaphor.diagram.general.simpleitem import Box, Ellipse, Line
 
 __all__ = [
-    "CommentItem",
-    "CommentLineItem",
     "DiagramItem",
     "Box",
     "Ellipse",
