+++ gaphor/diagram/general/tests/test_modelinglanguage.py
@@ -13,8 +13,6 @@ def test_lookup_element():
     modeling_language = GeneralModelingLanguage()
 
     assert modeling_language.lookup_element("DiagramItem")
-    assert modeling_language.lookup_element("CommentItem")
-    assert modeling_language.lookup_element("CommentLineItem")
     assert modeling_language.lookup_element("Box")
     assert modeling_language.lookup_element("Line")
     assert modeling_language.lookup_element("Ellipse")
