+++ gaphor/diagram/general/uicomponents.py
@@ -1,4 +1,3 @@
 # ruff: noqa: F401
 
-import gaphor.diagram.general.generaleditors
 import gaphor.diagram.general.generalpropertypages
