+++ gaphor/plugins/autolayout/tests/test_pydot.py
@@ -1,8 +1,6 @@
 import pytest
 
 from gaphor import UML
-from gaphor.core.modeling import Comment
-from gaphor.diagram.general.diagramitems import CommentItem, CommentLineItem
 from gaphor.diagram.tests.fixtures import connect
 from gaphor.plugins.autolayout.pydot import AutoLayout, parse_edge_pos, strip_quotes
 from gaphor.UML.diagramitems import (
@@ -15,6 +13,7 @@ from gaphor.UML.diagramitems import (
     ObjectFlowItem,
     PackageItem,
 )
+from gaphor.UML.general import CommentItem, CommentLineItem
 
 
 def test_layout_diagram(diagram, create):
@@ -49,7 +48,7 @@ def test_layout_with_comment(diagram, create, event_manager):
     connect(a, a.head, c1)
     connect(a, a.tail, c2)
 
-    comment = create(CommentItem, Comment)
+    comment = create(CommentItem, UML.Comment)
     comment_line = create(CommentLineItem)
     connect(comment_line, comment_line.head, comment)
     connect(comment_line, comment_line.tail, a)
