+++ gaphor/ui/tests/test_mainwindow.py
@@ -1,11 +1,12 @@
 import pytest
 
 from gaphor.application import Session
-from gaphor.core.modeling import Comment, Diagram
+from gaphor.core.modeling import Diagram
 from gaphor.diagram.event import DiagramOpened
 from gaphor.ui.abc import UIComponent
 from gaphor.ui.event import ElementOpened
 from gaphor.ui.tests.fixtures import iterate_until
+from gaphor.UML import Comment
 
 
 @pytest.fixture
